% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{theme_roboto}
\alias{theme_roboto}
\title{Minimal ggplot2 theme using the Roboto Condensed and Roboto Bold fonts}
\usage{
theme_roboto(...)
}
\arguments{
\item{...}{Arguments passed to \code{theme_minimal}}
}
\description{
Minimal ggplot2 theme using the Roboto Condensed and Roboto Bold fonts
}
\details{
The Roboto Condensed and Roboto Bold fonts are both Google fonts;
they can be found at \url{https://fonts.google.com/specimen/Roboto+Condensed}
and \url{https://fonts.google.com/specimen/Roboto}. These fonts must be
installed locally on your computer for this theme to work.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    theme_roboto()

}

